stm32mp-calibration: add clock calibration host
This commit adds what's necessary to build the application that will be
used as a daemon at runtime for HSI/CSI clock calibration.

It opens a session to the calibration PTA in optee_os and ask for a
calibration by invoking a calibration command. Then, it shuts down the
session and finalize the context.

Adds a timer that will invoke the calibration application every minute
after boot through a systemd service.

Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>
Change-Id: I70f613d8d2c736453e2df0de51830f478b41b8d7
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee-stm32mp-addons/+/241183